test(AES): Encryption service moved to a separate file and covered by tests
