+++ trade_portal/trade_portal/documents/views/oa.py
@@ -36,7 +36,7 @@ class OaCyphertextRetrieve(AllowCORSMixin, View):
         }
         # by default the ciphertext is base64-encoded
         if self.request.GET.get("key"):
-            from trade_portal.documents.services.lodge import AESCipher
+            from trade_portal.documents.services.encryption import AESCipher
 
             try:
                 cp = AESCipher(self.request.GET.get("key"))
